# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/user/auth": {
      "post": {
        "summary": "User Auth",
        "operationId": "user_auth_user_auth_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UserAuth"
                  }
                ],
                "title": "Data",
                "examples": [
                  {
                    "login": "tommy1884@mail.com",
                    "password_hash": "HashHashHashHash"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/registration/new": {
      "post": {
        "summary": "User Registration",
        "operationId": "user_registration_user_registration_new_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UserRegisterNew"
                  }
                ],
                "title": "Data",
                "examples": [
                  {
                    "login": "tommy1884@mail.com",
                    "password_hash": "HashHashHashHash",
                    "fullname": "Томас Шелби",
                    "avatar": "https://someimage.org/img.png",
                    "gender": 1,
                    "birthday": "20.01.1884",
                    "location_country": "Англия",
                    "location_city": "Смоллхит",
                    "location_region": "Бирмингем",
                    "salary": 20000000,
                    "skills": [
                      "Скачки",
                      "Гадание",
                      "Софтскиллс"
                    ],
                    "speciality": [
                      "Ведение Биснеса"
                    ]
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/registration/habr": {
      "post": {
        "summary": "User Registration",
        "operationId": "user_registration_user_registration_habr_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UserRegisterHabr"
                  }
                ],
                "title": "Data",
                "examples": [
                  {
                    "login": "tommy1884@mail.com",
                    "password_hash": "HashHashHashHash",
                    "habr_id": "tshelby"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/info": {
      "get": {
        "summary": "User Info",
        "operationId": "user_info_user_info_get",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/JustToken"
                  }
                ],
                "title": "Data",
                "examples": [
                  {
                    "token": "abcabcabc"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/update": {
      "post": {
        "summary": "User Update",
        "operationId": "user_update_user_update_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UserUpdate"
                  }
                ],
                "title": "Data",
                "examples": [
                  {
                    "token": "abcabcabc",
                    "password_hash": "HashHashHashHash",
                    "fullname": "Томас Шелби",
                    "avatar": "https://someimage.org/img.png",
                    "gender": 1,
                    "birthday": "20.01.1884",
                    "location_country": "Англия",
                    "location_city": "Смоллхит",
                    "location_region": "Бирмингем",
                    "salary": 20000000,
                    "skills": [
                      "Скачки",
                      "Гадание",
                      "Софтскиллс"
                    ],
                    "speciality": [
                      "Ведение Биснеса"
                    ]
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/messages/chats": {
      "get": {
        "summary": "User Messages Chats",
        "operationId": "user_messages_chats_user_messages_chats_get",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/JustToken"
                  }
                ],
                "title": "Data",
                "examples": [
                  {
                    "token": "abcabcabc"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/messages/send": {
      "post": {
        "summary": "User Messages Send",
        "operationId": "user_messages_send_user_messages_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UserSendMessage"
                  }
                ],
                "title": "Data",
                "examples": [
                  {
                    "token": "abcabcabc",
                    "chat_id": "Big_bro",
                    "message": "Артур, нужна твоя помощь. Приезжай в паб."
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/messages/refresh": {
      "get": {
        "summary": "User Messages Refresh",
        "operationId": "user_messages_refresh_user_messages_refresh_get",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UserRefreshMessages"
                  }
                ],
                "title": "Data",
                "examples": [
                  {
                    "token": "abcabcabc",
                    "chat_id": "Big_bro"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/person/info": {
      "get": {
        "summary": "Person Info",
        "operationId": "person_info_person_info_get",
        "parameters": [
          {
            "name": "person_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Person Id"
            },
            "examples": {
              "normal": {
                "summary": "Example with normal data",
                "value": "Big_bro"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/person/marks/add": {
      "post": {
        "summary": "Person Marks Add",
        "operationId": "person_marks_add_person_marks_add_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/PersonMark"
                  }
                ],
                "title": "Data",
                "examples": [
                  {
                    "token": "abcabcabc",
                    "person_id": "Big_bro"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/person/marks/status": {
      "get": {
        "summary": "Person Marks Status",
        "operationId": "person_marks_status_person_marks_status_get",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/PersonMark"
                  }
                ],
                "title": "Data",
                "examples": [
                  {
                    "token": "abcabcabc",
                    "person_id": "Big_bro"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/person/marks/remove": {
      "post": {
        "summary": "Person Marks Remove",
        "operationId": "person_marks_remove_person_marks_remove_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/PersonMark"
                  }
                ],
                "title": "Data",
                "examples": [
                  {
                    "token": "abcabcabc",
                    "person_id": "Big_bro"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/person/marks/all": {
      "get": {
        "summary": "Person Marks All",
        "operationId": "person_marks_all_person_marks_all_get",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/JustToken"
                  }
                ],
                "title": "Data",
                "examples": [
                  {
                    "token": "abcabcabc"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JustToken": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "JustToken"
      },
...
      "UserAuth": {
        "properties": {
          "login": {
            "type": "string",
            "title": "Login"
          },
          "password_hash": {
            "type": "string",
            "title": "Password Hash"
          }
        },
        "type": "object",
        "required": [
          "login",
          "password_hash"
        ],
        "title": "UserAuth"
      },
      "UserRefreshMessages": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "chat_id": {
            "type": "string",
            "title": "Chat Id"
          }
        },
        "type": "object",
        "required": [
          "token",
          "chat_id"
        ],
        "title": "UserRefreshMessages"
      },
      "UserRegisterHabr": {
        "properties": {
          "login": {
            "type": "string",
            "title": "Login"
          },
          "password_hash": {
            "type": "string",
            "title": "Password Hash"
          },
          "habr_id": {
            "type": "string",
            "title": "Habr Id"
          }
        },
        "type": "object",
        "required": [
          "login",
          "password_hash",
          "habr_id"
        ],
        "title": "UserRegisterHabr"
      },
      "UserRegisterNew": {
        "properties": {
          "login": {
            "type": "string",
            "title": "Login"
          },
          "password_hash": {
            "type": "string",
            "title": "Password Hash"
          },
          "fullname": {
            "type": "string",
            "title": "Fullname"
          },
          "avatar": {
            "type": "string",
            "title": "Avatar"
          },
          "gender": {
            "type": "integer",
            "title": "Gender"
          },
          "birthday": {
            "type": "string",
            "title": "Birthday"
          },
          "location_country": {
            "type": "string",
            "title": "Location Country"
          },
          "location_city": {
            "type": "string",
            "title": "Location City"
          },
          "location_region": {
            "type": "string",
            "title": "Location Region"
          },
          "salary": {
            "type": "integer",
            "title": "Salary"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills"
          },
          "speciality": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Speciality"
          }
        },
        "type": "object",
        "required": [
          "login",
          "password_hash",
          "fullname",
          "avatar",
          "gender",
          "birthday",
          "location_country",
          "location_city",
          "location_region",
          "salary",
          "skills",
          "speciality"
        ],
        "title": "UserRegisterNew"
      },
      "UserSendMessage": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "chat_id": {
            "type": "string",
            "title": "Chat Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "token",
          "chat_id",
          "message"
        ],
        "title": "UserSendMessage"
      },
      "UserUpdate": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "password_hash": {
            "type": "string",
            "title": "Password Hash"
          },
          "fullname": {
            "type": "string",
            "title": "Fullname"
          },
          "avatar": {
            "type": "string",
            "title": "Avatar"
          },
          "gender": {
            "type": "integer",
            "title": "Gender"
          },
          "birthday": {
            "type": "string",
            "title": "Birthday"
          },
          "location_country": {
            "type": "string",
            "title": "Location Country"
          },
          "location_city": {
            "type": "string",
            "title": "Location City"
          },
          "location_region": {
            "type": "string",
            "title": "Location Region"
          },
          "salary": {
            "type": "integer",
            "title": "Salary"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills"
          },
          "speciality": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Speciality"
          }
        },
        "type": "object",
        "required": [
          "token",
          "password_hash",
          "fullname",
          "avatar",
          "gender",
          "birthday",
          "location_country",
          "location_city",
          "location_region",
          "salary",
          "skills",
          "speciality"
        ],
        "title": "UserUpdate"
      },
...
    }
  }
}